feat: Disable `@typescript-eslint/no-non-null-assertion`
…but warn about confusing uses.

Rationale:

When using `noUncheckedIndexedAccess`, the compiler will repeatedly
warn you, and a lot of the time adding `!` to the end of the expression
is the most ergonomical way to fix it.

This is also about consistency. The non-null assertion operator is
just a sugar/shorthand for `as Exclude<T, null | undefined>`, and such
expressions do not trigger warnings.