feat: Raise `@typescript-eslint/no-explicit-any` back to `error` level
Explicitly using `any` (outside `...rest:args[]` function arguments
is in vast majority of cases a massive red flag for a codebase.

A good linting config ought to treat it as "error" by default, and
force developers to make informed decisions — either to improve the
type, or at least be explicit about making an exception via a
`eslint-disable` directive.